backend/requirements.txt
# Core LLM and RAG
langchain>=0.3.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0
langchain-community>=0.3.0

# Vector Database
qdrant-client>=1.12.0

# Embeddings
sentence-transformers>=3.0.0
langchain-huggingface>=0.1.0

# FastAPI Backend
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.20
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Document Processing
pypdf>=5.1.0
python-docx>=1.1.2
python-magic-bin>=0.4.14; sys_platform == "win32"
python-magic>=0.4.27; sys_platform != "win32"
docx2txt>=0.8

# Azure Communication Services (Teams integration)
azure-communication-callautomation>=1.3.0
azure-communication-identity>=1.5.0

# Speech Services (free, no API keys needed)
vosk>=0.3.44
gTTS>=2.5.0

# Utilities
python-dotenv>=1.0.1
websockets>=12.0

